Fix compiler warning.

+++ prov/zhpe/include/zhpe.h
@@ -325,7 +325,7 @@ extern int zhpe_eq_def_sz;
 extern char *zhpe_pe_affinity_str;
 extern char *zhpe_fab_backend_prov;
 extern char *zhpe_fab_backend_dom;
-extern int zhpe_ep_max_eager_sz;
+extern size_t zhpe_ep_max_eager_sz;
 
 static inline void *zhpe_mremap(void *old_address, size_t old_size,
 				size_t new_size)
